+++ gaia-medication/App.tsx
@@ -1,0 +1,25 @@
+import { StatusBar } from 'expo-status-bar';
+import { StyleSheet, Text, View } from 'react-native';
+import Home from './.expo/Home';
+import Navigation from './.expo/Navigation';
+import { NavigationContainer } from '@react-navigation/native';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+
+
+export default function App() {
+  return (
+    <View style={styles.container}>
+      <Navigation />
+    </View>
+    
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    //alignItems: 'center',
+   // justifyContent: 'center',
+  },
+});
